backend/pyproject.toml
[project]
name = "hyb8nate"
version = "0.1.0-alpha"
description = "Automatically scale Kubernetes deployments up/down based on time schedules"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "mainmain-rl" }
]
license = { text = "MIT" }
keywords = ["kubernetes", "scheduler", "devops", "automation", "finops", "deployment"]

dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "sqlalchemy>=2.0.36",
    "kubernetes>=31.0.0",
    "pydantic[email]>=2.10.3",
    "pydantic-settings>=2.6.1",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.1",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    "pytz>=2025.2",
    "loguru>=0.7.3",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.black]
line-length = 120
target-version = ["py313"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[flake8]
max-line-length = 120
